gratitude_journal_service/requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
Django==3.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-graphql-jwt==0.3.0
fastdiff==0.2.0
graphene==2.1.8
graphene-django==2.7.1
graphql-core==2.2.1
graphql-relay==2.0.1
promise==2.2.1
psycopg2-binary==2.8.4
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
redgreenunittest==0.1.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
snapshottest==0.5.1
sqlparse==0.3.0
termcolor==1.1.0
wasmer==0.3.0
